graph: backend: dnnl: exec: gated_mlp uses unique_ptr instead of raw ptr
